+++ src/main/java/de/dealog/msg/event/MessageEventPayloadConverter.java
@@ -1,9 +1,13 @@
 package de.dealog.msg.event;
 
 import com.google.common.base.Converter;
+import de.dealog.common.model.MessageEventPayload;
 import de.dealog.msg.persistence.Message;
 import de.dealog.msg.persistence.MessageEntity;
-import de.dealog.common.model.MessageEventPayload;
+import org.geolatte.geom.Polygon;
+import org.geolatte.geom.codec.Wkt;
+import org.geolatte.geom.codec.WktDecoder;
+import org.geolatte.geom.crs.CoordinateReferenceSystems;
 
 import javax.inject.Singleton;
 
@@ -19,6 +23,9 @@ public class MessageEventPayloadConverter extends Converter<MessageEventPayload,
         message.setIdentifier(payload.getIdentifier());
         message.setHeadline(payload.getHeadline());
         message.setDescription(payload.getDescription());
+        WktDecoder decoder = Wkt.newDecoder(Wkt.Dialect.POSTGIS_EWKT_1);
+        Polygon<?> polygon = (Polygon<?>) decoder.decode(payload.getGeocode(), CoordinateReferenceSystems.WGS84);
+        message.setGeocode(polygon);
         return message;
     }
 
@@ -28,6 +35,7 @@ public class MessageEventPayloadConverter extends Converter<MessageEventPayload,
         payload.setIdentifier(message.getIdentifier());
         payload.setHeadline(message.getHeadline());
         payload.setDescription(message.getDescription());
+        payload.setGeocode(message.getGeocode().toString());
         return payload;
     }
 }