+++ src/main/java/de/dealog/msg/service/MessageService.java
@@ -4,8 +4,11 @@ import de.dealog.msg.persistence.Message;
 import de.dealog.msg.persistence.MessageEntity;
 import de.dealog.msg.persistence.MessageRepository;
 import io.quarkus.hibernate.orm.panache.PanacheQuery;
+import io.quarkus.panache.common.Parameters;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.geolatte.geom.G2D;
+import org.geolatte.geom.Point;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
@@ -28,9 +31,19 @@ public class MessageService {
      * @param size
      * @return
      */
-    public PagedList<? extends Message> list(final int page, final int size) {
+    public PagedList<? extends Message> list(final Point<G2D> point,
+                                             final int page, final int size) {
         log.debug("List messages for page {} and size {} ...", page, size);
-        PanacheQuery<MessageEntity> messageQuery = messageRepository.findAll();
+        PanacheQuery<MessageEntity> messageQuery;
+        if (point != null) {
+
+            messageQuery = MessageEntity.find(
+                    "within(:point, geocode) = true",
+                    Parameters.with("point", point));
+        } else {
+            messageQuery = messageRepository.findAll();
+        }
+
         List<MessageEntity> list = messageQuery.page(page, size).list();
         long count = messageQuery.count();
         int pageCount = messageQuery.pageCount();
@@ -51,7 +64,7 @@ public class MessageService {
      * @param message
      */
     public void create(final Message message) {
-        log.debug("Create new message {}" , message);
+        log.debug("Create message {}" , message);
         if (messageRepository.findByIdentifier(message.getIdentifier()) == null) {
             messageRepository.persistAndFlush((MessageEntity) message);
         } else {